Changes Gfn
Hello, sorry for the delay ... I was very busy at work. Unfortunately not even from act 5 I spent in poe kkkk
I added a few lines, but I didn't deal with the fact that there was no text to be read, as it would change the heart of the code. I just changed what you couldn't test because you can't use geforce now, using the hotkeys you can already activate the macro, but you still need a refinement for this case where we don't have the txt. Anyway thanks for the attention, and if you need anything, give me a touch.